refactor(server): extract orchestration service from routes + chat backend
Both packages/server/src/routes/orchestration.ts and
packages/server/src/ai/backend.ts implemented the same 4 operations
(readyNodes/claimNode/releaseNode/conflictScan) with duplicated logic.
Maintenance hazard — any change had to be made in two places.

Extracts to packages/server/src/services/orchestration.ts:
  - 4 service functions with one canonical implementation each
  - buildIsDonePredicate / buildInProgressIds / buildTodoIds helpers
  - ClaimOwnershipError + OrchestrationNotFoundError typed exceptions
    so the HTTP route can map them to status codes while the chat
    backend just lets them propagate.

Routes shrink from 376 to 121 lines. AI chat backend shrinks by 141
lines. Net -387 lines across the two consumers; ~280 lines in the
new service file. Total reduction ~100 lines.

Addresses #111 review (#116).
